requirements.txt
Django==1.8
http://www.stickpeople.com/projects/python/win-psycopg/2.6.0/psycopg2-2.6.0.win-amd64-py2.7-pg9.4.1-release.exe
django-allauth
python-dateutil
pillow
django-crispy-forms